Ensure consistent naming of presets
