+++ src/validators/query.ts
@@ -106,10 +106,7 @@ export const WithInstructionSchema = z.record(
 );
 
 // Including Instructions.
-export const IncludingInstructionSchema = z.union([
-  z.literal('all'),
-  z.array(z.string()),
-]);
+export const IncludingInstructionSchema = z.record(z.lazy((): z.ZodTypeAny => GetQuerySchema));
 
 // Ordering Instructions.
 export const OrderedByInstructionSchema = z
