+++ tests/errors.test.ts
@@ -313,76 +313,4 @@ test('get multiple records with empty `before` instruction', () => {
   expect(error).toHaveProperty('code', 'MISSING_INSTRUCTION');
 });
 
-test('get single record including parent schemas (one-to-many) without shortcut', () => {
-  const queries: Array<Query> = [
-    {
-      get: {
-        account: {
-          including: ['members'],
-        },
-      },
-    },
-  ];
 
-  const schemas: Array<Schema> = [
-    {
-      slug: 'account',
-    },
-    {
-      slug: 'member',
-    },
-  ];
-
-  let error: Error | undefined;
-
-  try {
-    compileQueries(queries, schemas);
-  } catch (err) {
-    error = err as Error;
-  }
-
-  expect(error).toBeInstanceOf(RoninError);
-  expect(error).toHaveProperty(
-    'message',
-    'The provided `including` shortcut "members" does not exist in schema "Account".',
-  );
-  expect(error).toHaveProperty('code', 'INVALID_INCLUDING_VALUE');
-});
-
-test('get single record for pre-defined match without shortcut', () => {
-  const queries: Array<Query> = [
-    {
-      get: {
-        account: {
-          for: {
-            'active-member': 'acc_39h8fhe98hefah8',
-          },
-        },
-      },
-    },
-  ];
-
-  const schemas: Array<Schema> = [
-    {
-      slug: 'account',
-    },
-    {
-      slug: 'member',
-    },
-  ];
-
-  let error: Error | undefined;
-
-  try {
-    compileQueries(queries, schemas);
-  } catch (err) {
-    error = err as Error;
-  }
-
-  expect(error).toBeInstanceOf(RoninError);
-  expect(error).toHaveProperty(
-    'message',
-    'Preset "active-member" does not exist in schema "Account".',
-  );
-  expect(error).toHaveProperty('code', 'PRESET_NOT_FOUND');
-});
